# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Confidential Document Analyst API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/hybrid": {
      "post": {
        "summary": "Hybrid Endpoint",
        "operationId": "hybrid_endpoint_api_hybrid_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HybridRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "summary": "History",
        "operationId": "history_api_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/cases": {
      "get": {
        "summary": "List Cases Endpoint",
        "operationId": "list_cases_endpoint_api_cases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Case"
                  },
                  "type": "array",
                  "title": "Response List Cases Endpoint Api Cases Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Case Endpoint",
        "operationId": "create_case_endpoint_api_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}": {
      "get": {
        "summary": "Get Case Endpoint",
        "operationId": "get_case_endpoint_api_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Case Endpoint",
        "operationId": "update_case_endpoint_api_cases__case_id__patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Case Endpoint",
        "operationId": "delete_case_endpoint_api_cases__case_id__delete",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/files": {
      "post": {
        "summary": "Upload File Endpoint",
        "operationId": "upload_file_endpoint_api_cases__case_id__files_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_endpoint_api_cases__case_id__files_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Files Endpoint",
        "operationId": "list_files_endpoint_api_cases__case_id__files_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseFile"
                  },
                  "title": "Response List Files Endpoint Api Cases  Case Id  Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/analyze": {
      "post": {
        "summary": "Analyze Endpoint",
        "operationId": "analyze_endpoint_api_cases__case_id__analyze_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStartedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/progress": {
      "get": {
        "summary": "Progress Endpoint",
        "operationId": "progress_endpoint_api_cases__case_id__progress_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisProgress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/findings": {
      "get": {
        "summary": "Findings Endpoint",
        "operationId": "findings_endpoint_api_cases__case_id__findings_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Finding"
                  },
                  "title": "Response Findings Endpoint Api Cases  Case Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/report": {
      "get": {
        "summary": "Report Endpoint",
        "operationId": "report_endpoint_api_cases__case_id__report_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/report/export": {
      "get": {
        "summary": "Export Report Endpoint",
        "operationId": "export_report_endpoint_api_cases__case_id__report_export_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/{full_path}": {
      "get": {
        "summary": "Spa Fallback",
        "description": "Serve index.html for all non-API routes (SPA client-side routing).",
        "operationId": "spa_fallback__full_path__get",
        "parameters": [
          {
            "name": "full_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Full Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisProgress": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "total",
          "completed",
          "status"
        ],
        "title": "AnalysisProgress"
      },
      "AnalysisStartedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "case_id"
        ],
        "title": "AnalysisStartedResponse"
      },
      "Body_upload_file_endpoint_api_cases__case_id__files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_endpoint_api_cases__case_id__files_post"
      },
      "Case": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus",
            "default": "open"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "default": "low"
          },
          "cloud_consent": {
            "type": "boolean",
            "title": "Cloud Consent",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CaseFile"
            },
            "type": "array",
            "title": "Files"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "Case"
      },
      "CaseFile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "format": {
            "$ref": "#/components/schemas/FileFormat"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          },
          "preview": {
            "type": "string",
            "title": "Preview",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "filename",
          "format",
          "created_at"
        ],
        "title": "CaseFile"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "open",
          "analyzing",
          "complete"
        ],
        "title": "CaseStatus"
      },
      "CreateCaseRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateCaseRequest"
      },
      "DeleteResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "DeleteResponse"
      },
      "FileFormat": {
        "type": "string",
        "enum": [
          "whatsapp",
          "imessage",
          "email",
          "plain_text"
        ],
        "title": "FileFormat"
      },
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "category": {
            "$ref": "#/components/schemas/FindingCategory"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "local"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "file_id",
          "category",
          "severity",
          "quote",
          "explanation",
          "created_at"
        ],
        "title": "Finding"
      },
      "FindingCategory": {
        "type": "string",
        "enum": [
          "threat",
          "scam",
          "abuse",
          "pattern",
          "timeline_event",
          "communication"
        ],
        "title": "FindingCategory"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "status",
          "models"
        ],
        "title": "HealthResponse"
      },
      "HybridRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "HybridRequest"
      },
      "HybridResponse": {
        "properties": {
          "function_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Function Calls"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "total_time": {
            "type": "number",
            "title": "Total Time"
          }
        },
        "type": "object",
        "required": [
          "function_calls",
          "source",
          "confidence",
          "total_time"
        ],
        "title": "HybridResponse"
      },
      "Report": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "case_name": {
            "type": "string",
            "title": "Case Name"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntry"
            },
            "type": "array",
            "title": "Timeline"
          },
          "findings_by_category": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Finding"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Findings By Category"
          },
          "stats": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "type": "object",
            "title": "Stats"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "case_name",
          "generated_at",
          "summary",
          "risk_level",
          "timeline",
          "findings_by_category",
          "stats"
        ],
        "title": "Report"
      },
      "RiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel"
      },
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity"
      },
      "TimelineEntry": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          }
        },
        "type": "object",
        "required": [
          "date",
          "event",
          "severity",
          "source_file",
          "quote"
        ],
        "title": "TimelineEntry"
      },
      "UpdateCaseRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "cloud_consent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Consent"
          }
        },
        "type": "object",
        "title": "UpdateCaseRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}